Fix for rainlab.blog

+++ updates/extend_rainlab_blog_posts_table.php
@@ -20,9 +20,11 @@ class ExtendRainlabBlogPostsTable extends Migration
     public function down()
     {
         if (PluginManager::instance()->hasPlugin('RainLab.Blog')) {
-            Schema::table('rainlab_blog_posts', function ($table) {
-                $table->dropColumn('arcane_seo_options');
-            });
+            if (Schema::hasColumn('rainlab_blog_posts', 'arcane_seo_options')) {
+                Schema::table('rainlab_blog_posts', function ($table) {
+                    $table->dropColumn('arcane_seo_options');
+                });
+            }
         }
     }
 }
